+++ src/session-impl.test.ts
@@ -868,6 +868,17 @@ test("SessionImpl.getOutbox()", async (t) => {
       "https://example.com/ap/note/01941f29-7c00-7fe8-ab0a-7b593990a3c0",
     ]);
   });
+
+  await t.test("messages should have update and delete methods", async () => {
+    const outbox = session.getOutbox({ order: "oldest" });
+    const messages = await Array.fromAsync(outbox);
+    assert.strictEqual(messages.length, 4);
+
+    for (const message of messages) {
+      assert.strictEqual(typeof message.update, "function");
+      assert.strictEqual(typeof message.delete, "function");
+    }
+  });
 });
 
 export interface SentActivity {